Move `StaticScheduler` to a separate module.
